requirements.txt
numpy==1.22.1
opencv_python==4.5.5.62
Pillow==9.0.0
